+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - support for type `Uint64` with `JsonObject` and `JsonArray`.
+- methods `FromJson` and `ToJson` to transform data two-way.
 
 ------------------------